fix(js): prevent dep cycle between store and components
