+++ js/packages/components/helpers.ts
@@ -86,15 +86,6 @@ export const pbDateToNum = (pbTimestamp?: number | Long | string | null): number
 	}
 }
 
-export const getRandomColor = () => {
-	const letters = '0123456789ABCDEF'
-	let color = '#'
-	for (let i = 0; i < 6; i++) {
-		color += letters[Math.floor(Math.random() * 16)]
-	}
-	return color
-}
-
 export const showNeedRestartNotification = (showNotification: any, ctx: MsgrState, t: any) => {
 	showNotification({
 		title: t('notification.need-restart.title'),
@@ -125,21 +116,4 @@ export const readProfileNotification = async (
 			[type]: numberNotifs - 1,
 		},
 	})
-}
-
-export const randomizeThemeColor = () => {
-	return {
-		'main-text': getRandomColor(),
-		'main-background': getRandomColor(),
-		'secondary-text': getRandomColor(),
-		'background-header': getRandomColor(),
-		'secondary-background-header': getRandomColor(),
-		'alt-secondary-background-header': getRandomColor(),
-		'reverted-main-text': getRandomColor(),
-		'positive-asset': getRandomColor(),
-		'negative-asset': getRandomColor(),
-		'warning-asset': getRandomColor(),
-		'input-background': getRandomColor(),
-		shadow: getRandomColor(),
-	}
 }
