+++ js/packages/store/context.ts
@@ -6,10 +6,10 @@ import { ServiceClientType } from '@berty-tech/grpc-bridge/welsh-clients.gen'
 import { globals } from '@berty-tech/config'
 import { Service } from '@berty-tech/grpc-bridge'
 import rpcBridge from '@berty-tech/grpc-bridge/rpc/rpc.bridge'
+import { randomizeThemeColor } from '@berty-tech/styles'
 import defaultTheme from '@berty-tech/styles/colors.json'
 import pinkTheme from '@berty-tech/styles/pinktheme-default.json'
 import darkTheme from '@berty-tech/styles/darktheme-default.json'
-import { randomizeThemeColor } from '@berty-tech/components/helpers'
 
 import { ParsedInteraction } from './types.gen'
 import { SoundKey } from './sounds'
