+++ js/packages/styles/index.tsx
@@ -1,3 +1,4 @@
 export { useStyles, Provider, Consumer, ctx } from './hook'
 export type { ColorsTypes, ColorsStyles } from './types'
 export { defaultStylesDeclaration } from './map-declaration'
+export * from './helpers'
